UPDATE: working hexagonal system
* finish hexagonal coordinate system + draw one hexagon
* drawing is only temporary -> TODO: create proper drawing system
* prob. delete HexPainter? Use draw component?
* CartesianPosition is model -> TODO: move it
* create algorithm filling the board with hexagons

+++ src/model/units/Unit.java
@@ -35,6 +35,6 @@ public abstract class Unit extends JComponent {
 	@Override
 	protected void paintComponent(Graphics g) {
 		super.paintComponent(g);
-		graphicModel.getUnitImage().paintIcon(this, g, position.getX(), position.getY());
+		graphicModel.getUnitImage().paintIcon(this, g, (int) position.getX(), (int) position.getY());
 	}
 }
